refactor: 오류가 메시지를 가지게 수정
