+++ src/main/java/exception/DuplicateNameException.java
@@ -3,6 +3,6 @@ package exception;
 public class DuplicateNameException extends IllegalArgumentException {
 
     public DuplicateNameException() {
-        super();
+        super("중복되는 이름이 존재합니다. 다시 입력해주세요");
     }
 }
