+++ src/test/java/view/OutputViewTest.java
@@ -7,7 +7,6 @@ import java.io.PrintStream;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import view.output.OutputView;
 
 class OutputViewTest {
